+++ packages/core-database-postgres/lib/queries/migrations/create.sql
@@ -1,0 +1,1 @@
+INSERT INTO migrations (id, name) VALUES (${id}, ${name})