+++ packages/core-database-postgres/lib/queries/migrations/find.sql
@@ -1,0 +1,3 @@
+SELECT *
+FROM migrations
+WHERE name = ${name}